split oracle connmanager.
